Reorganized structure of auditor.inbound
