+++ auditor.inbound/src/main/java/module-info.java
@@ -6,6 +6,8 @@
 module auditor.inbound {
   exports io.github.tkaczenko.auditor.inbound;
   exports io.github.tkaczenko.auditor.inbound.config;
+    exports io.github.tkaczenko.auditor.inbound.internal.service;
+  exports io.github.tkaczenko.auditor.inbound.api;
 
   requires auditor.core;
   requires static lombok;
