+++ demo/src/main/java/io/github/tkaczenko/auditor/demo/service/filter/DemoControllerAuditRequestFilter.java
@@ -4,8 +4,8 @@ import io.github.tkaczenko.auditor.core.internal.service.AuditDateTimeProvider;
 import io.github.tkaczenko.auditor.core.internal.AuditFacade;
 import io.github.tkaczenko.auditor.core.internal.factory.body.BodyHttpReaderService;
 import io.github.tkaczenko.auditor.core.internal.factory.headers.HeadersHttpReaderService;
-import io.github.tkaczenko.auditor.inbound.AuditInboundRequestFilter;
-import io.github.tkaczenko.auditor.inbound.IpAddressSupplier;
+import io.github.tkaczenko.auditor.inbound.api.AuditInboundRequestFilter;
+import io.github.tkaczenko.auditor.inbound.internal.service.IpAddressSupplier;
 import org.springframework.stereotype.Component;
 
 @Component
